backend/requirements.txt
flask
flask-cors
pandas
sqlalchemy
scikit-learn
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
